+++ src/blas.rs
@@ -1,7 +1,11 @@
+mod gemv;
+mod gerc;
 mod iamax;
 mod nrm2;
 mod scal;
 
+pub(crate) use gemv::{gemv, gemv_transpose};
+pub(crate) use gerc::gerc;
 pub(crate) use iamax::iamax;
 pub(crate) use nrm2::nrm2;
 pub(crate) use scal::{rscal, scal};