+++ src/lapack/larfg.rs
@@ -8,20 +8,16 @@ where
     A::Real: Real,
     S: DataMut<Elem = A>,
 {
-    if x.is_empty() {
-        return (alpha.re(), x, A::zero());
-    }
-
     let mut x_norm = blas::nrm2(&x);
-    if x_norm == A::zero().re() {
+    if x_norm == A::zero().re() && alpha.im() == A::zero().re() {
         return (alpha.re(), x, A::zero());
     }
 
     let mut beta = -lapack::lapy3(alpha.re(), alpha.im(), x_norm).copysign(alpha.re());
     let safe_min = A::Real::sfmin() / A::Real::eps();
-    let safe_min_recip = safe_min.recip();
     let mut knt = 0;
     if beta.abs() < safe_min {
+        let safe_min_recip = safe_min.recip();
         loop {
             knt += 1;
             blas::rscal(safe_min_recip, &mut x);
@@ -46,6 +42,7 @@ where
 mod test {
     use approx::{assert_abs_diff_eq, AbsDiffEq};
     use ndarray::arr1;
+    use num_complex::Complex64;
 
     #[test]
     fn larfg_empty() {
@@ -57,11 +54,24 @@ mod test {
     }
 
     #[test]
-    fn larfg() {
+    fn larfg_real() {
         let x = arr1(&[2., 3., 2.]);
         let (beta, x, tau) = super::larfg(1., x);
         assert_abs_diff_eq!(beta, -4.24264069, epsilon = 1e-6);
         assert!(x.abs_diff_eq(&arr1(&[0.38148714, 0.57223071, 0.38148714]), 1e-6));
         assert_abs_diff_eq!(tau, 1.23570226, epsilon = 1e-6);
+    }
+
+    #[test]
+    fn larfg_complex() {
+        let x = arr1(&[Complex64::new(2., 3.), Complex64::new(3., 2.)]);
+        let (beta, x, tau) = super::larfg(Complex64::new(1., 1.), x);
+        assert_abs_diff_eq!(beta, -5.2915026221291805, epsilon = 1e-8);
+        assert_abs_diff_eq!(x[0].re, 0.38397859, epsilon = 1e-6);
+        assert_abs_diff_eq!(x[0].im, 0.41580232, epsilon = 1e-6);
+        assert_abs_diff_eq!(x[1].re, 0.51436575, epsilon = 1e-6);
+        assert_abs_diff_eq!(x[1].im, 0.23613346, epsilon = 1e-6);
+        assert_abs_diff_eq!(tau.re, 1.1889822365046137, epsilon = 1e-8);
+        assert_abs_diff_eq!(tau.im, 0.18898223650461363, epsilon = 1e-8);
     }
 }